requirements.txt
#python3.6.7

django==2.2
django-summernote==0.8.11.4
django-solo==1.1.3